// src/apis/models/FireBytes_Unified_Wave_Pacs_PacsActionPlanGroupSettings_Dtos_GetPacsActionPlanGroupSettingForEditOutput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FireBytes_Unified_Wave_Pacs_PacsActionPlanGroupSettings_Dtos_PacsActionPlanGroupSettingEditDto } from './FireBytes_Unified_Wave_Pacs_PacsActionPlanGroupSettings_Dtos_PacsActionPlanGroupSettingEditDto';
export type FireBytes_Unified_Wave_Pacs_PacsActionPlanGroupSettings_Dtos_GetPacsActionPlanGroupSettingForEditOutput = {
    pacsActionPlanGroupSettingList: Array<FireBytes_Unified_Wave_Pacs_PacsActionPlanGroupSettings_Dtos_PacsActionPlanGroupSettingEditDto>;
};