// src/apis/models/FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderPhones_PacsCardHolderPhoneEditDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_Pacs_PacsCardHolders_Dtos_PacsCardHolderPhones_PacsCardHolderPhoneEditDto = {
    id?: string | null;
    pacsCardHolderId?: string | null;
    phoneType?: string | null;
    phoneNumber?: string | null;
};